+++ nautilus_trader/accounting/accounts/betting.pyx
@@ -18,7 +18,7 @@ from decimal import Decimal
 from nautilus_trader.accounting.accounts.cash cimport CashAccount
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.rust.enums cimport AccountType
-from nautilus_trader.model.c_enums.order_side cimport OrderSide
+from nautilus_trader.core.rust.enums cimport OrderSide
 from nautilus_trader.model.data.bet cimport Bet
 from nautilus_trader.model.data.bet cimport nautilus_to_bet
 from nautilus_trader.model.instruments.base cimport Instrument