+++ nautilus_trader/trading/strategy.pyx
@@ -47,6 +47,7 @@ from nautilus_trader.common.timer cimport TimeEvent
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.message cimport Event
 from nautilus_trader.core.rust.enums cimport oms_type_from_str
+from nautilus_trader.core.rust.enums cimport order_side_to_str
 from nautilus_trader.core.uuid cimport UUID4
 from nautilus_trader.execution.algorithm cimport ExecAlgorithmSpecification
 from nautilus_trader.execution.messages cimport CancelAllOrders
@@ -56,7 +57,6 @@ from nautilus_trader.execution.messages cimport QueryOrder
 from nautilus_trader.execution.messages cimport SubmitOrder
 from nautilus_trader.execution.messages cimport SubmitOrderList
 from nautilus_trader.indicators.base.indicator cimport Indicator
-from nautilus_trader.model.c_enums.order_side cimport OrderSideParser
 from nautilus_trader.model.c_enums.position_side cimport PositionSideParser
 from nautilus_trader.model.c_enums.time_in_force cimport TimeInForce
 from nautilus_trader.model.data.bar cimport Bar
@@ -741,7 +741,7 @@ cdef class Strategy(Actor):
     cpdef void cancel_all_orders(
         self,
         InstrumentId instrument_id,
-        OrderSide order_side = OrderSide.NONE,
+        OrderSide order_side = OrderSide.NO_ORDER_SIDE,
         ClientId client_id = None,
     ) except *:
         """
@@ -775,7 +775,7 @@ cdef class Strategy(Actor):
             side=order_side,
         )
 
-        cdef str order_side_str = " " + OrderSideParser.to_str(order_side) if order_side != OrderSide.NONE else ""
+        cdef str order_side_str = " " + order_side_to_str(order_side) if order_side != OrderSide.NO_ORDER_SIDE else ""
         if not open_orders and not emulated_orders:
             self.log.info(
                 f"No open or emulated{order_side_str} "